:truck: move page to create access request from accounts to core app
