+++ src/zac/core/urls.py
@@ -12,6 +12,7 @@ from .views.processes import (
 )
 from .views.search import SearchIndexView, SearchView
 from .views.zaken import (
+    AccessRequestCreateView,
     FetchZaakObjecten,
     Index,
     ZaakAccessRequestsView,
@@ -48,6 +49,11 @@ urlpatterns = [
                     "<bronorganisatie>/<identificatie>/afhandelen/",
                     ZaakAfhandelView.as_view(),
                     name="zaak-afhandeling",
+                ),
+                path(
+                    "<bronorganisatie>/<identificatie>/access-requests/add/",
+                    AccessRequestCreateView.as_view(),
+                    name="access-request-create",
                 ),
                 path(
                     "<bronorganisatie>/<identificatie>/access-requests/",
