+++ src/zac/core/views/zaken.py
@@ -3,7 +3,7 @@ from typing import Any, Dict, List, Optional
 
 from django.contrib.auth.mixins import LoginRequiredMixin
 from django.db import transaction
-from django.urls import reverse
+from django.urls import reverse, reverse_lazy
 from django.views.generic import FormView, TemplateView
 
 from extra_views import ModelFormSetView
@@ -25,7 +25,12 @@ from zac.contrib.kownsl.data import ReviewRequest
 from zac.utils.api_models import serialize
 
 from ..base_views import BaseDetailView, BaseListView, SingleObjectMixin
-from ..forms import AccessRequestHandleForm, ZaakAfhandelForm, ZakenFilterForm
+from ..forms import (
+    AccessRequestCreateForm,
+    AccessRequestHandleForm,
+    ZaakAfhandelForm,
+    ZakenFilterForm,
+)
 from ..permissions import zaken_close, zaken_inzien, zaken_set_result
 from ..services import (
     find_zaak,
@@ -305,6 +310,32 @@ class ZaakActiviteitenView(PermissionRequiredMixin, BaseDetailView):
         return zaak
 
 
+class AccessRequestCreateView(LoginRequiredMixin, FormView):
+    form_class = AccessRequestCreateForm
+    template_name = "core/create_access_request.html"
+    #  todo add thanks page?
+    success_url = reverse_lazy("core:index")
+
+    def get_zaak(self):
+        return find_zaak(**self.request.resolver_match.kwargs)
+
+    def get_form_kwargs(self):
+        kwargs = super().get_form_kwargs()
+
+        kwargs["requester"] = self.request.user
+        kwargs["zaak"] = self.get_zaak()
+        return kwargs
+
+    def get_context_data(self, **kwargs):
+        context = super().get_context_data(**kwargs)
+        context.update({"zaak": self.get_zaak()})
+        return context
+
+    def form_valid(self, form):
+        form.save()
+        return super().form_valid(form)
+
+
 class ZaakAccessRequestsView(LoginRequiredMixin, ModelFormSetView):
     form_class = AccessRequestHandleForm
     template_name = "core/zaak_access_requests.html"
@@ -328,7 +359,7 @@ class ZaakAccessRequestsView(LoginRequiredMixin, ModelFormSetView):
         context.update({"zaak": self.get_zaak()})
         return context
 
-    # @transaction.atomic
+    @transaction.atomic
     def formset_valid(self, formset):
         response = super().formset_valid(formset)
